credentialsAreValid in backend, no security breach
